Amanda Jiménez Lobato exam 12/12/2021
The exam and a picture of what I intended to obtain with the code of the first approach to the Voronoi regions. I did not actually got to truncate or make disappear the lines once they intersected another line, in order to obtain the polygons, but it would have looked like that if I had got to make it.